- improve loading on server - remove logs
